[update]カテゴリーID,サブカテゴリーID/データ型bigint変更

+++ db/migrate/20230314011052_add_category_id_and_subcategory_id_to_post_images.rb
@@ -1,7 +1,7 @@
 class AddCategoryIdAndSubcategoryIdToPostImages < ActiveRecord::Migration[6.1]
   def change
-    add_column :post_images, :category_id, :integer
-    add_column :post_images, :sub_category_id, :integer
+    add_column :post_images, :category_id, :bigint
+    add_column :post_images, :sub_category_id, :bigint
     add_foreign_key :post_images, :categories, column: :category_id
     add_foreign_key :post_images, :sub_categories, column: :sub_category_id
   end
